LyapunovExponent.jl Documentation

This file contains the implementation of functions related to the computation of Lyapunov Exponents and Kaplan-Yorke dimension.

Lyapunov Exponent

We implement two different algorithms to compute the Lyapunov exponent. The first implementation lyapunovExponent requires numerical integration of the model with unperturbed initial conditions and perturbed initial conditions to compute the Lyapunov exponent. The second implementation lyapunovExponentJacobian leverages the tangent map or Jacobian.

LiftAndLearn.ChaosGizmo.lyapunovExponent — Function

lyapunovExponent(ops::operators, integrator::Function, ic::AbstractArray, 
        options::LE_options, params...) → λ::Vector or (λ::Vector, λ_all::Matrix)

Compute the Lyapunov exponents of a dynamical system using the algorirhtm in [edson2019] which uses the algorithm from [benettin1980] and [shimada1979] to compute them.

Arguments

ops::operators: operators for the full model

integrator::Function: integrator for the model

ic::AbstractArray: initial condition

options::LE_options: options for the algorithm

params...: keyword arguments for the integrator

Returns

λ::Vector: the Lyapunov exponents

λ_all::Matrix: all Lyapunov exponents if options.history is true

Integrator function format

The integrator function must be in the form of integrator(ops, tspan, ic) where ops is the operator struct, tspan is the time span to integrate over, and ic is the initial condition. The integrator must return the solution at the final time. And the integrator can have additional parameters as keyword arguments params....

Example integrator function for the Lorenz system using the 4th order Runge-Kutta method:

function lorenz_integrator(ops::LnL.operators, tspan::AbstractArray, IC::Array; params...)
    K = length(tspan)
    N = size(IC,1)
    f = let A = ops.A, H = ops.H
        (x, t) -> A*x + H*kron(x,x)
    end
    xk = zeros(N,K)
    xk[:,1] = IC

    for k in 2:K
        timestep = tspan[k] - tspan[k-1]
        k1 = f(xk[:,k-1], tspan[k-1])
        k2 = f(xk[:,k-1] + 0.5 * timestep * k1, tspan[k-1] + 0.5 * timestep)
        k3 = f(xk[:,k-1] + 0.5 * timestep * k2, tspan[k-1] + 0.5 * timestep)
        k4 = f(xk[:,k-1] + timestep * k3, tspan[k-1] + timestep)
        xk[:,k] = xk[:,k-1] + (timestep / 6.0) * (k1 + 2*k2 + 2*k3 + k4)
    end
    return xk
end

References

[edson2019] R. A. Edson, J. E. Bunder, T. W. Mattner, and A. J. Roberts, “Lyapunov Exponents of the Kuramoto–Sivashinsky PDE,” The ANZIAM Journal, vol. 61, no. 3, pp. 270–285, Jul. 2019, doi: 10.1017/S1446181119000105.

[benettin1980] G. Benettin, L. Galgani, A. Giorgilli, and J.-M. Strelcyn, “Lyapunov Characteristic Exponents for smooth dynamical systems and for hamiltonian systems; a method for computing all of them. Part 1: Theory,” Meccanica, vol. 15, no. 1, pp. 9–20, Mar. 1980, doi: 10.1007/BF02128236.

[shimada1979] [1] I. Shimada and T. Nagashima, “A Numerical Approach to Ergodic Problem of Dissipative Dynamical Systems,” Progress of Theoretical Physics, vol. 61, no. 6, pp. 1605–1616, Jun. 1979, doi: 10.1143/PTP.61.1605.

lyapunovExponent(ops::operators, integrator::Function, Vr::AbstractArray, 
    ic::AbstractArray, options::LE_options, params...) → λ::Vector or (λ::Vector, λ_all::Matrix)

Compute the Lyapunov exponents of a dynamical system using the algorirhtm in [edson2019] for the reduced model. This is a dispatched function for the case when the reduced model is used.

Arguments

model::Abstract_Models: the dynamical model

ops::operators: operators for the reduced model

integrator::Function: integrator for the model

Vr::AbstractArray: basis for the reduced model

ic::AbstractArray: initial condition

options::LE_options: options for the algorithm

params...: keyword arguments for the integrator

Returns

λ::Vector: the Lyapunov exponents

λ_all::Matrix: all Lyapunov exponents if options.history is true

Integrator function format

The integrator function must be in the form of integrator(ops, tspan, ic) where ops is the operator struct, tspan is the time span to integrate over, and ic is the initial condition. The integrator must return the solution at the final time. And the integrator can have additional parameters as keyword arguments params....

LiftAndLearn.ChaosGizmo.lyapunovExponentJacobian — Function

lyapunovExponentJacobian(ops::operators, integrator::Function, jacobian::Function, 
    ic::AbstractArray, options::LE_options, params...) → λ::Vector or (λ::Vector, λ_all::Matrix)

Compute the Lyapunov exponents of the full order dynamical system using the Jacobian of the system. Refer to References for more details.

Arguments

ops::operators: operators for the full/reduced model

integrator::Function: integrator for the full/reduced model

jacobian::Function: Jacobian of the system

ic::AbstractArray: initial condition

options::LE_options: options for the algorithm

params...: keyword arguments for the integrator

Returns

λ::Vector: the Lyapunov exponents

λ_all::Matrix: all Lyapunov exponents if options.history is true

Integrator function format

The integrator function must be in the form of integrator(ops, tspan, ic) where ops is the operator struct, tspan is the time span to integrate over, and ic is the initial condition. The integrator must return the solution at the final time. And the integrator can have additional parameters as keyword arguments params....

Function to generate the Jacobian

The Jacobian function must be in the form of jacobian(ops, x) where ops is the operator struct and x is the state. This Jacobian matrix constructs the Linear Tangent Map (LTM) of the system. The Jacobian for a linear-quadratic system

\[\dot{\mathbf{x}} = \mathbf{Ax} + \mathbf{H}(\mathbf{x} \otimes \mathbf{x})\]

is for exmaple

function jacobian(ops, x)
    A = ops.A
    H = ops.H
    return A + H * kron(1.0I(n), x) + H * kron(x, 1.0I(n))
end

References

GitHub code by ni-sha-c

GitHub code by ThomasSavary08

P. V. Kuptsov and U. Parlitz, “Theory and computation of covariant Lyapunov vectors,” arXiv, 2011, doi: 10.48550/ARXIV.1105.5228.

lyapunovExponentJacobian(ops::operators, integrator::Function, jacobian::Function, Vr::AbstractArray, 
    ic::AbstractArray, options::LE_options, params...) → λ::Vector or (λ::Vector, λ_all::Matrix)

Compute the Lyapunov exponents of a reduced order dynamical system using the Jacobian of the system. Refer to References for more details.

Arguments

ops::operators: operators for the full/reduced model

integrator::Function: integrator for the full/reduced model

jacobian::Function: Jacobian of the system

Vr::AbstractArray: basis for the reduced model

ic::AbstractArray: initial condition

options::LE_options: options for the algorithm

params...: keyword arguments for the integrator

Returns

λ::Vector: the Lyapunov exponents

λ_all::Matrix: all Lyapunov exponents if options.history is true

Integrator function format

The integrator function must be in the form of integrator(ops, tspan, ic) where ops is the operator struct, tspan is the time span to integrate over, and ic is the initial condition. The integrator must return the solution at the final time. And the integrator can have additional parameters as keyword arguments params....

Function to generate the Jacobian

The Jacobian function must be in the form of jacobian(ops, x) where ops is the operator struct and x is the state. This Jacobian matrix constructs the Linear Tangent Map (LTM) of the system. The Jacobian for a linear-quadratic system

\[\dot{\mathbf{x}} = \mathbf{Ax} + \mathbf{H}(\mathbf{x} \otimes \mathbf{x})\]

is for exmaple

function jacobian(ops, x)
    A = ops.A
    H = ops.H
    return A + H * kron(1.0I(n), x) + H * kron(x, 1.0I(n))
end

References

GitHub code by ni-sha-c

GitHub code by ThomasSavary08

P. V. Kuptsov and U. Parlitz, “Theory and computation of covariant Lyapunov vectors,” arXiv, 2011, doi: 10.48550/ARXIV.1105.5228.

Kaplan-Yorke dimension

The Kaplan-Yorke dimension represents the effective dimension of the chaotic attractor. For more details see this explanation. In ChaosGizmo, the Kaplan-Yorke dimension is computed for given Lyapunov exponents.

LiftAndLearn.ChaosGizmo.kaplanYorkeDim — Function

kaplanYorkeDim(λs::AbstractVector; sorted::Bool=true) → Dky::Float64

Compute the Kaplan-Yorke dimension from the Lyapunov exponents.

Arguments

λs::AbstractVector{<:Real}: the Lyapunov exponents

sorted::Bool: whether the Lyapunov exponents are sorted in descending order

Returns

Float64: the Kaplan-Yorke dimension

Example

julia> using LiftAndLearn.ChaosGizmo

julia> CG = LiftAndLearn.ChaosGizmo

julia> lyapunovSpectrum = [0.5, 0.1, -0.2, -0.4]  # Example spectrum

julia> kyDimension = CG.kaplanYorkeDim(lyapunovSpectrum)

Implementation Details

Integration step of the tangent map

To integrate the tangent map/Jacobian further in time we employ multiple integration schemes ranging from Euler method, 4th order Runge-Kutta, and strong stability preserving Runge-Kutta (SSPRK3).

LiftAndLearn.ChaosGizmo.EULER — Function

EULER(J, Q, dt) → Qnew

Euler method for the perturbation integrator.

Arguments

J::AbstractMatrix: Jacobian of the system

Q::AbstractArray: perturbation state

dt::Real: timestep

Returns

Qnew::AbstractArray: updated perturbation state

LiftAndLearn.ChaosGizmo.RK2 — Function

RK2(J, Q, dt) → Qnew

2nd order Runge-Kutta method for the perturbation integrator.

Arguments

J::AbstractMatrix: Jacobian of the system

Q::AbstractArray: perturbation state

dt::Real: timestep

Returns

Qnew::AbstractArray: updated perturbation state

LiftAndLearn.ChaosGizmo.RK4 — Function

RK4(J, Q, dt) → Qnew

4th order Runge-Kutta method for the perturbation integrator.

Arguments

J::AbstractMatrix: Jacobian of the system

Q::AbstractArray: perturbation state

dt::Real: timestep

Returns

Qnew::AbstractArray: updated perturbation state

LiftAndLearn.ChaosGizmo.RALSTON4 — Function

RALSTON4(J, Q, dt) → Qnew

Ralston's fourth-order method for the perturbation integrator.

Arguments

J::AbstractMatrix: Jacobian of the system

Q::AbstractArray: perturbation state

dt::Real: timestep

Returns

Qnew::AbstractArray: updated perturbation state

LiftAndLearn.ChaosGizmo.SSPRK3 — Function

SSPRK3(J, Q, dt) → Qnew

3rd order Strong Stability Preserving Runge-Kutta method for the perturbation integrator.

Arguments

J::AbstractMatrix: Jacobian of the system

Q::AbstractArray: perturbation state

dt::Real: timestep

Returns

Qnew::AbstractArray: updated perturbation state

Adjusting the QR decomposition

In every step of computing the Lyapunov exponent, it is necessary to reorthogonalize the tangent map. And, in this process we must adjust the signs in the Q and R matrices using the following routine:

LiftAndLearn.ChaosGizmo.posDiag_QR! — Function

posDiag_QR!(Q::AbstractMatrix, R::AbstractMatrix)

Make sure the diagonal of the R matrix is positive.

Arguments

Q::AbstractMatrix: orthogonal directions

R::AbstractMatrix: upper triangular matrix

Returns

nothing

Options for Lyapunov exponent computation

To compute the Lyapunov exponents, the user must define the LE_options with necessary parameters. It is crucial to tune the parameters based on your system at hand.

LiftAndLearn.ChaosGizmo.LE_options — Type

struct LE_options

Options for Lypuanov exponents.

Fields

m::Integer: the number of the most exponents to compute

τ::Real: time to simulate the system before computing exponents

Δt::Real: timestep for the integrator

Δτ::Real: timestep for the integrator (before computing exponents)

T::Real: time between reorthogonalization steps

N::Integer: the total number of reorthogonalization steps

ϵ::Real: perturbation magnitude

verbose::Bool: print progress to stdout

history::Bool: return all LEs

pert_integrator::Function: integrator for the perturbed system

Default values

m = 1

τ = 1000

Δt = 1e-2

Δτ = Δt

T = Δt

N = 1000

ϵ = 1e-6

verbose = false

history = false

pert_integrator = nothing

Perturbation integrator

The perturbation integrator must be in the form of integrator(J, Q, dt) where J is the Jacobian of the system, Q is the perturbation state, and dt is the timestep. The integrator must return the updated perturbation state. The default integrator is the 4th order Runge-Kutta method:'

function RK4(J, Q, dt)
    # RK4 steps
    k1 = J * Q
    k2 = J * (Q + 0.5*dt*k1)
    k3 = J * (Q + 0.5*dt*k2)
    k4 = J * (Q + dt*k3)
    # Update the perturbation state
    Qnew = Q + (dt/6) * (k1 + 2*k2 + 2*k3 + k4)
    return Qnew
end

You can select the following explicit method integrators available in the ChaosGizmo submodule:

EULER(J, Q, dt): Euler method

RK2(J, Q, dt): 2nd order Runge-Kutta method

RK4(J, Q, dt): 4th order Runge-Kutta method

SSPRK3(J, Q, dt): 3rd order Strong Stability Preserving Runge-Kutta method

RALSTON4(J, Q, dt): Ralston's fourth-order method

Note

If you are using the Jacobian/Linear Tangent Map (LTM) method, T should be set equal to Δt (i.e. Δt = T) or T = 2*Δt.

Example

We use a 9-dimensional Lorenz system to analyze the Lyapunov exponents and Kaplan-Yorke dimension of a reduced model. This model is introduced in Reiterer et al.. We reduce this system to an order of 7 from 9 for the analysis. Here are some cool visualizations of the model's dynamics.

<img src="assets/images/ChaosGizmo/9dimlorenz3d.svg" width="300" align="right"> <img src="assets/images/ChaosGizmo/9dimlorenzframe1.svg" width="300" align="right"> <img src="assets/images/ChaosGizmo/9dimlorenzframe2.svg" width="300" align="right"> <img src="assets/images/ChaosGizmo/9dimlorenzframe3.svg" width="300" align="right">

using LinearAlgebra: svd, I
using LiftAndLearn
const LnL = LiftAndLearn

# Define the 9-dimensional Lorenz system
a = 0.5  # wave number in the horizontal direction
b1 = 4*(1+a^2) / (1+2*a^2)
b2 = (1+2*a^2) / (2*(1+a^2))
b3 = 2*(1-a^2) / (1+a^2)
b4 = a^2 / (1+a^2)
b5 = 8*a^2 / (1+2*a^2)
b6 = 4 / (1+2*a^2)

sigma = 0.5  # Prandtl number
r = 15.10  # reduced Rayleigh number

n = 9  # dimension of the system
# Linear operator
A = zeros(n,n)
A[1,1] = -sigma * b1
A[1,7] = -sigma * b2
A[2,2] = -sigma
A[2,9] = -sigma / 2
A[3,3] = -sigma * b1
A[3,8] = sigma * b2
A[4,4] = -sigma
A[4,9] = sigma / 2
A[5,5] = -sigma * b5
A[6,6] = -b6
A[7,1] = -r
A[7,7] = -b1
A[8,3] = r
A[8,8] = -b1
A[9,2] = -r
A[9,4] = r
A[9,9] = -1

# Quadratic operator
indices = [
    (2,4,1), (4,4,1), (3,5,1),
    (1,4,2), (2,5,2), (4,5,2),
    (2,4,3), (2,2,3), (1,5,3),
    (2,3,4), (2,5,4), (4,5,4),
    (2,2,5), (4,4,5),
    (2,9,6), (4,9,6),
    (5,8,7), (4,9,7),
    (5,7,8), (2,9,8),
    (2,6,9), (4,6,9), (4,7,9), (2,8,9)
]
values = [
    -1, b4, b3,
    1, -1, 1,
    1, -b4, -b3,
    -1, -1, 1,
    0.5, -0.5,
    1, -1,
    2, -1,
    -2, 1,
    -2, 2, 1, -1
]
H = LnL.makeQuadOp(n, indices, values; which_quad_term="H")
F = LnL.makeQuadOp(n, indices, values; which_quad_term="F")
lorenz9_ops = LnL.operators(A=A, H=H, F=F)

# Define some helper functions
function lorenz_jacobian(ops::LnL.operators, x::AbstractArray)
    n = size(x,1)
    return ops.A + ops.H * kron(I(n),x) + ops.H*kron(x,I(n))
end

function lorenz_integrator(ops::LnL.operators, tspan::AbstractArray, IC::Array; params...)
    K = length(tspan)
    N = size(IC,1)
    f = let A = ops.A, H = ops.H, F = ops.F
        if ops.H == 0
            (x, t) -> A*x + F*LnL.vech(x*x')
        else
            (x, t) -> A*x + H*kron(x,x)
        end
    end
    xk = zeros(N,K)
    xk[:,1] = IC

    for k in 2:K
        timestep = tspan[k] - tspan[k-1]
        k1 = f(xk[:,k-1], tspan[k-1])
        k2 = f(xk[:,k-1] + 0.5 * timestep * k1, tspan[k-1] + 0.5 * timestep)
        k3 = f(xk[:,k-1] + 0.5 * timestep * k2, tspan[k-1] + 0.5 * timestep)
        k4 = f(xk[:,k-1] + timestep * k3, tspan[k-1] + timestep)
        xk[:,k] = xk[:,k-1] + (timestep / 6.0) * (k1 + 2*k2 + 2*k3 + k4)
    end
    return xk
end

# Then construct the intrusive POD-Galerkin reduced model
# Lyapunov exponent of POD reduced model
data1 = lorenz_integrator(lorenz9_ops, 0:1e-2:1e3, 2*rand(9).-1)
data2 = lorenz_integrator(lorenz9_ops, 0:1e-2:1e3, 2*rand(9).-1)
data3 = lorenz_integrator(lorenz9_ops, 0:1e-2:1e3, 2*rand(9).-1)
data4 = lorenz_integrator(lorenz9_ops, 0:1e-2:1e3, 2*rand(9).-1)
data5 = lorenz_integrator(lorenz9_ops, 0:1e-2:1e3, 2*rand(9).-1)
data = hcat(data1, data2, data3, data4, data5)
rmax = 7
Vr = svd(data).U[:,1:rmax]   # choose rmax columns
rom_option = LnL.LS_options(
    system=LnL.sys_struct(is_lin=true, is_quad=true),
)
oprom = LnL.intrusiveMR(lorenz9_ops, Vr, rom_option)

# Now, we can compute the Lyapunov spectrum and the Kaplan-Yorke dimension using the method without the Tangent map
const CG = LnL.ChaosGizmo
x0 = [0.01, 0, 0.01, 0.0, 0.0, 0.0, 0, 0, 0.01]
options = CG.LE_options(N=1e4, τ=1e2, τ0=0.0, Δt=1e-2, m=rmax, T=1e-2, verbose=false, history=true)
λr1, λr1_all = CG.lyapunovExponent(oprom, lorenz_integrator, Vr, x0, options)
dkyr1 = CG.kaplanYorkeDim(λr1; sorted=false)
println("Kaplan-Yorke dimension (without Jacobian): ", dkyr1)

using Plots
p = plot()
for i in 1:rmax
    plot!(p, λr1_all[i,:], label="λ$i", lw=2)
end
plot!(p, xlabel="reorthonormalization steps", ylabel="Lyapunov exponent", legend=:right, fontfamily="Computer Modern")
display(p)

# You can also compute them using the method with the Tangent map.
λr2, λr2_all = CG.lyapunovExponentJacobian(oprom, lorenz_integrator, lorenz_jacobian, Vr' * x0, options)
dkyr2 = CG.kaplanYorkeDim(λr2; sorted=false)
println("Kaplan-Yorke dimension (with Jacobian): ", dkyr2)

p = plot()
for i in 1:rmax
    plot!(p, λr2_all[i,:], label="λ$i", lw=2)
end
plot!(p, xlabel="reorthonormalization steps", ylabel="Lyapunov exponent", legend=:right, fontfamily="Computer Modern")
display(p)

Kaplan-Yorke dimension (without Jacobian): 2.063988293798444
GKS: cannot open display - headless operation mode active
Kaplan-Yorke dimension (with Jacobian): 1.512420398877544