Lifting

This page gives an explanation for the mathematical concept of lifting and provides example code for its implementation.

Lift Map Structure

LiftAndLearn.lifting — Type

struct lifting

Lifting map structure.

Fields

N: number of variables of the original nonlinear dynamics

Nl: number of variables of the lifted system

lift_funcs: array of lifting transformation functions

map: function to map the data to the new mapped states including original states

mapNL: function to map the data to only the additional lifted states

Tip

For a simple pendulum we have

\[\begin{bmatrix} \dot{x}_1 \\ \dot{x}_2 \end{bmatrix} = \begin{bmatrix} x_2 \\ -\frac{g}{l} \sin(x_1) \end{bmatrix}\]

The lifted system becomes

\[\begin{bmatrix} \dot{x}_1 \\ \dot{x}_2 \\ \dot{x}_3 \\ \dot{x}_4 \end{bmatrix} = \begin{bmatrix} x_2 \\ -\frac{g}{l} x_3 \\ x_2 x_4 \\ -x_2 x_3 \end{bmatrix}\]

when $x_3 = \sin(x_1)$ and $x_4 = \cos(x_1)$. Which if coded, would look like this:

lifter = LnL.lifting(2, 4, [x -> sin.(x[1]), x -> cos.(x[1])])

Construct Lifted POD Basis from Data

LiftAndLearn.liftedBasis — Function

liftedBasis(W, Nl, gp, ro) → Vr

Create the block-diagonal POD basis for the new lifted system data

Arguments

w: lifted data matrix

Nl: number of variables of the lifted state dynamics

gp: number of grid points for each variable

ro: vector of the reduced orders for each basis

Return

Vr: block diagonal POD basis

An example implementation would be:

using LiftAndLearn
using Random
LnL = LiftAndLearn
W = round.(rand(30,100), digits=4)
LnL.liftedBasis(W, 3, 10, [2,3,4])

30×9 BlockDiagonals.BlockDiagonal{Float64, Matrix{Float64}}:
 -0.32925    0.0975642  0.0  0.0  0.0  …   0.0         0.0         0.0
 -0.325521   0.0513447  0.0  0.0  0.0      0.0         0.0         0.0
 -0.290275   0.439847   0.0  0.0  0.0      0.0         0.0         0.0
 -0.361881  -0.149806   0.0  0.0  0.0      0.0         0.0         0.0
 -0.32109   -0.578482   0.0  0.0  0.0      0.0         0.0         0.0
 -0.297477  -0.105549   0.0  0.0  0.0  …   0.0         0.0         0.0
 -0.298083   0.355633   0.0  0.0  0.0      0.0         0.0         0.0
 -0.314197  -0.273568   0.0  0.0  0.0      0.0         0.0         0.0
 -0.325397   0.422722   0.0  0.0  0.0      0.0         0.0         0.0
 -0.292174  -0.214818   0.0  0.0  0.0      0.0         0.0         0.0
  ⋮                                    ⋱                          
  0.0        0.0        0.0  0.0  0.0     -0.517591   -0.0629654   0.294798
  0.0        0.0        0.0  0.0  0.0     -0.374601   -0.243872   -0.220103
  0.0        0.0        0.0  0.0  0.0     -0.0594075   0.0542056  -0.635892
  0.0        0.0        0.0  0.0  0.0      0.102588    0.367897   -0.0707006
  0.0        0.0        0.0  0.0  0.0  …   0.518445   -0.268716   -0.275386
  0.0        0.0        0.0  0.0  0.0      0.286947    0.622152    0.172846
  0.0        0.0        0.0  0.0  0.0     -0.158079   -0.155638    0.266424
  0.0        0.0        0.0  0.0  0.0     -0.110401    0.310782    0.245624
  0.0        0.0        0.0  0.0  0.0     -0.121988   -0.049735   -0.180975