Viscous Burgers' Equation

LiftAndLearn.Burgers — Module

Viscous Burgers' equation model

LiftAndLearn.Burgers.Abstract_Models — Type

Abstract_Models

Abstract type for the models.

LiftAndLearn.Burgers.burgers — Type

mutable struct burgers <: LiftAndLearn.Burgers.Abstract_Models

Viscous Burgers' equation model

\[\frac{\partial u}{\partial t} = \mu\frac{\partial^2 u}{\partial x^2} - u\frac{\partial u}{\partial x}\]

Fields

Omega::Vector{Float64}: spatial domain

T::Vector{Float64}: temporal domain

D::Vector{Float64}: parameter domain

Δx::Float64: spatial grid size

Δt::Float64: temporal step size

IC::VecOrMat{Float64}: initial condition

x::Vector{Float64}: spatial grid points

t::Vector{Float64}: temporal points

μs::Union{Vector{Float64},Float64}: parameter vector

Xdim::Int64: spatial dimension

Tdim::Int64: temporal dimension

Pdim::Int64: parameter dimension

BC::String: boundary condition

generateABFmatrix::Function: function to generate A, B, F matrices

generateMatrix_NC_periodic::Function: function to generate A, F matrices for the non-energy preserving Burgers' equation. (Non-conservative Periodic boundary condition)

generateMatrix_C_periodic::Function: function to generate A, F matrices for the non-energy preserving Burgers' equation. (conservative periodic boundary condition)

generateEPmatrix::Function: function to generate A, F matrices for the Burgers' equation. (Energy-preserving form)

semiImplicitEuler::Function: function to integrate the system using semi-implicit Euler scheme

LiftAndLearn.Burgers.burgers — Method

burgers(Omega, T, D, Δx, Δt, Pdim, BC) → burgers

Constructor for the Burgers' equation model.

LiftAndLearn.Burgers.generateABFmatrix — Method

generateABFmatrix(model, μ) → A, B, F

Generate A, B, F matrices for the Burgers' equation.

Arguments

model: Burgers' equation model

μ: parameter value

Returns

A: A matrix

B: B matrix

F: F matrix

LiftAndLearn.Burgers.generateEPmatrix — Method

generateEPmatrix(model, μ) → A, F

Generate A, F matrices for the Burgers' equation. (Energy-preserving form)

Arguments

model: Burgers' equation model

μ: parameter value

Returns

A: A matrix

F: F matrix

LiftAndLearn.Burgers.generateMatrix_C_periodic — Method

generateMatrix_C_periodic(model, μ) → A, F

Generate A, F matrices for the non-energy preserving Burgers' equation. (conservative periodic boundary condition)

Arguments

model: Burgers' equation model

μ: parameter value

Returns

A: A matrix

F: F matrix

LiftAndLearn.Burgers.generateMatrix_NC_periodic — Method

generateMatrix_NC_periodic(model, μ) → A, F

Generate A, F matrices for the non-energy preserving Burgers' equation. (Non-conservative Periodic boundary condition)

Arguments

model: Burgers' equation model

μ: parameter value

Returns

A: A matrix

F: F matrix

LiftAndLearn.Burgers.semiImplicitEuler — Method

semiImplicitEuler(A, F, tdata, IC) → states

Semi-Implicit Euler scheme without control (dispatch)

Arguments

A: linear state operator

F: quadratic state operator

tdata: time data

IC: initial condtions

Returns

states: integrated states

LiftAndLearn.Burgers.semiImplicitEuler — Method

semiImplicitEuler(A, B, F, U, tdata, IC) → states

Semi-Implicit Euler scheme

Arguments

A: linear state operator

B: linear input operator

F: quadratic state operator

U: input data

tdata: time data

IC: initial condtions

Returns

states: integrated states

LiftAndLearn.Burgers.vech — Method

vech(A) → v

Half-vectorization operation

Arguments

A: matrix to half-vectorize

Returns

v: half-vectorized form