Fitzhugh-Nagumo Equation

LiftAndLearn.FitzHughNagumo — Module

FitzHugh-Nagumo PDE model

LiftAndLearn.FitzHughNagumo.FitzHughNagumoModel — Type

mutable struct FitzHughNagumoModel <: LiftAndLearn.AbstractModel

FitzHugh-Nagumo PDE model

\[\begin{aligned} \frac{\partial u}{\partial t} &= \epsilon^2\frac{\partial^2 u}{\partial x^2} + u(u-0.1)(1-u) - v + g \\ \frac{\partial v}{\partial t} &= hu + \gamma v + g \end{aligned}\]

where $u$ and $v$ are the state variables, $g$ is the control input, and $h$, $\gamma$, and $\epsilon$ are the parameters. Specifically, for this problem we assume the control input to begin

\[g(t) = \alpha t^3 \exp(-\beta t)\]

where $\alpha$ and $\beta$ are the parameters that are going to be varied for training.

Fields

spatial_domain::Tuple{Real,Real}: spatial domain

time_domain::Tuple{Real,Real}: temporal domain

alpha_input_param_domain::Tuple{Real,Real}: parameter domain

beta_input_param_domain::Tuple{Real,Real}: parameter domain

Δx::Real: spatial grid size

Δt::Real: temporal step size

BC::Symbol: boundary condition

IC::Array{Float64}: initial condition

IC_lift::Array{Float64}: initial condition

xspan::Vector{Float64}: spatial grid points

tspan::Vector{Float64}: temporal points

alpha_input_params::Union{Real,AbstractArray{<:Real}}: input parameter vector

beta_input_params::Union{Real,AbstractArray{<:Real}}: input parameter vector

spatial_dim::Int64: spatial dimension

time_dim::Int64: temporal dimension

param_dim::Dict{Symbol,<:Int}: parameter dimension

full_order_model::Function: full order model

lifted_finite_diff_model::Function: lifted finite difference model

LiftAndLearn.FitzHughNagumo.FitzHughNagumoModel — Method

FitzHughNagumoModel(;spatial_domain::Tuple{Real,Real}, time_domain::Tuple{Real,Real}, Δx::Real, Δt::Real, 
                alpha_input_params::Union{AbstractArray{<:Real},Real}, beta_input_params::Union{AbstractArray{<:Real},Real}, 
                BC::Symbol=:neumann) → FitzHughNagumoModel

Constructor FitzHugh-Nagumo PDE model

LiftAndLearn.FitzHughNagumo.full_order_model — Method

full_order_model(k, l) → A, B, C, K, f

Create the full order operators with the nonlinear operator expressed as f(x).

Arguments

k::Int64: number of spatial grid points

l::Float64: spatial domain length

Returns

A::SparseMatrixCSC{Float64,Int64}: A matrix

B::SparseMatrixCSC{Float64,Int64}: B matrix

C::SparseMatrixCSC{Float64,Int64}: C matrix

K::SparseMatrixCSC{Float64,Int64}: K matrix

f::Function: nonlinear operator

LiftAndLearn.FitzHughNagumo.lifted_finite_diff_model — Method

lifted_finite_diff_model(k, l) → A, B, C, H, N, K

Generate the full order operators used for the intrusive model operators

Arguments

k::Int64: number of spatial grid points

l::Float64: spatial domain length

Returns

A::SparseMatrixCSC{Float64,Int64}: A matrix

B::SparseMatrixCSC{Float64,Int64}: B matrix

C::SparseMatrixCSC{Float64,Int64}: C matrix

H::SparseMatrixCSC{Float64,Int64}: H matrix

N::SparseMatrixCSC{Float64,Int64}: N matrix

K::SparseMatrixCSC{Float64,Int64}: K matrix