Fisher-KPP Equation

LiftAndLearn.FisherKPP — Module

Fisher Kolmogorov-Petrovsky-Piskunov equation PDE model

LiftAndLearn.FisherKPP.FisherKPPModel — Type

mutable struct FisherKPPModel <: LiftAndLearn.AbstractModel

Fisher Kolmogorov-Petrovsky-Piskunov equation (Fisher-KPP) model is a reaction-diffusion equation or logistic diffusion process in population dynamics. The model is given by the following PDE:

\[\frac{\partial u}{\partial t} = D\frac{\partial^2 u}{\partial x^2} + ru(1-u)\]

where $u$ is the state variable, $D$ is the diffusion coefficient, and $r$ is the growth rate.

Fields

spatial_domain::Tuple{Real,Real}: spatial domain

time_domain::Tuple{Real,Real}: temporal domain

diffusion_coeff_domain::Tuple{Real,Real}: parameter domain (diffusion coeff)

growth_rate_domain::Tuple{Real,Real}: parameter domain (growth rate)

Δx::Real: spatial grid size

Δt::Real: temporal step size

xspan::Vector{<:Real}: spatial grid points

tspan::Vector{<:Real}: temporal points

spatial_dim::Int: spatial dimension

time_dim::Int: temporal dimension

diffusion_coeffs::Union{AbstractArray{<:Real},Real}: diffusion coefficient

growth_rates::Union{AbstractArray{<:Real},Real}: growth rate

param_dim::Dict{Symbol,<:Int}: parameter dimension

IC::AbstractArray{<:Real}: initial condition

BC::Symbol: boundary condition

finite_diff_model::Function: model using Finite Difference

integrate_model::Function: integrator using Crank-Nicholson (linear) Explicit (nonlinear) method

LiftAndLearn.FisherKPP.finite_diff_mixed_model — Method

finite_diff_mixed_model(model::fisherkpp, D::Real, r::Real)

Create the matrices A (linear operator), B (input operator), and F (quadratic operator) for the Fisher-KPP model using the mixed boundary condition. If the spatial domain is [0,1], then we assume u(0,t) to be homogeneous dirichlet boundary condition and u(1,t) to be Neumann boundary condition of some function h(t).

Arguments

N::Real: spatial dimension

Δx::Real: spatial grid size

D::Real: diffusion coefficients

r::Real: growth rates

Returns

A::SparseMatrixCSC{Float64,Int}: linear operator

B::SparseMatrixCSC{Float64,Int}: input operator

F::SparseMatrixCSC{Float64,Int}: quadratic operator

LiftAndLearn.FisherKPP.finite_diff_model — Method

finite_diff_model(model::fisherkpp, D::Real, r::Real)

Create the matrices A (linear operator) and F (quadratic operator) for the Fisher-KPP model. For different boundary conditions, the matrices are created differently.

Arguments

model::FisherKPPModel: Fisher-KPP model

D::Real: diffusion coefficients

r::Real: growth rates

LiftAndLearn.FisherKPP.finite_diff_periodic_model — Method

finite_diff_model(model::fisherkpp, D::Real, r::Real)

Create the matrices A (linear operator) and F (quadratic operator) for the Fisher-KPP model.

Arguments

N::Real: spatial dimension

Δx::Real: spatial grid size

D::Real: diffusion coefficients

r::Real: growth rates

Returns

A::SparseMatrixCSC{Float64,Int}: linear operator

F::SparseMatrixCSC{Float64,Int}: quadratic operator

LiftAndLearn.FisherKPP.integrate_model — Method

integrate_model(A::AbstractArray{<:Real}, F::AbstractArray{<:Real}, IC::AbstractArray{<:Real}, 
                tspan::AbstractArray{<:Real}; const_stepsize::Bool=false)

Integrate the Fisher-KPP model using the Crank-Nicholson (linear) Explicit (nonlinear) method.

Arguments

A::AbstractArray{<:Real}: linear operator

F::AbstractArray{<:Real}: quadratic operator

IC::AbstractArray{<:Real}: initial condition

tspan::AbstractArray{<:Real}: time span

const_stepsize::Bool=false: constant time step size

Returns

u::AbstractArray{<:Real}: solution

LiftAndLearn.FisherKPP.integrate_model — Method

integrate_model(A::AbstractArray{T}, B::AbstractArray{T}, F::AbstractArray{T}, U::AbstractArray{T}, 
                tspan::AbstractArray{T}, IC::AbstractArray{T}; const_stepsize::Bool=false) where T<:Real

Integrate the Fisher-KPP model using the Crank-Nicholson (linear) Adam-Bashforth (nonlinear) method.

Arguments

A::AbstractArray{T}: linear operator

B::AbstractArray{T}: input operator

F::AbstractArray{T}: quadratic operator

U::AbstractArray{T}: input

tspan::AbstractArray{T}: time span

IC::AbstractArray{T}: initial condition

const_stepsize::Bool=false: constant time step size

Returns

state::AbstractArray{T}: solution