Kuramoto-Sivashinsky Equation

LiftAndLearn.KuramotoSivashinsky — Module

Kuramoto-Sivashinsky equation PDE model

LiftAndLearn.KuramotoSivashinsky.KuramotoSivashinskyModel — Type

mutable struct KuramotoSivashinskyModel <: LiftAndLearn.AbstractModel

Kuramoto-Sivashinsky equation PDE model

\[\frac{\partial u}{\partial t} = -\mu\frac{\partial^4 u}{\partial x^4} - \frac{\partial^2 u}{\partial x^2} - u\frac{\partial u}{\partial x}\]

where $u$ is the state variable and $\mu$ is the viscosity coefficient.

Fields

spatial_domain::Tuple{Real,Real}: spatial domain

time_domain::Tuple{Real,Real}: temporal domain

param_domain::Tuple{Real,Real}: parameter domain

Δx::Real: spatial grid size

Δt::Real: temporal step size

BC::Symbol: boundary condition

IC::Array{Float64}: initial condition

xspan::Vector{Float64}: spatial grid points

tspan::Vector{Float64}: temporal points

diffusion_coeffs::Union{Real,AbstractArray{<:Real}}: parameter vector

fourier_modes::Vector{Float64}: Fourier modes

spatial_dim::Int64: spatial dimension

time_dim::Int64: temporal dimension

param_dim::Int64: parameter dimension

conservation_type::Symbol: conservation type

model_type::Symbol: model type

finite_diff_model::Function: finite difference model

pseudo_spectral_model::Function: pseudo spectral model

elementwise_pseudo_spectral_model::Function: element-wise pseudo spectral model

spectral_galerkin_model::Function: spectral Galerkin model

integrate_model::Function: integrator

jacobian::Function: Jacobian matrix

LiftAndLearn.KuramotoSivashinsky.KuramotoSivashinskyModel — Method

KuramotoSivashinskyModel(
;
    spatial_domain,
    time_domain,
    Δx,
    Δt,
    diffusion_coeffs,
    BC,
    conservation_type,
    model_type
)

Constructor for the Kuramoto-Sivashinsky equation model.

LiftAndLearn.KuramotoSivashinsky.elementwise_pseudo_spectral_model — Method

elementwise_pseudo_spectral_model(model, μ) → A, F

Generate A, F matrices for the Kuramoto-Sivashinsky equation using the Fast Fourier Transform method (element-wise).

Arguments

model: Kuramoto-Sivashinsky equation model

μ: parameter value

Returns

A: A matrix

F: F matrix

LiftAndLearn.KuramotoSivashinsky.integrate_elementwise_pseudo_spectral_model — Method

integrate_elementwise_pseudo_spectral_model(A, F, tdata, IC)

Integrator using Crank-Nicholson Adams-Bashforth method for (FFT) (element-wise)

Arguments

A: A matrix

F: F matrix

tdata: temporal points

IC: initial condition

Returns

u: state matrix

uhat: state matrix in the Fourier space

LiftAndLearn.KuramotoSivashinsky.integrate_finite_diff_model — Method

integrate_finite_diff_model(
    A,
    F,
    tdata,
    IC;
    const_stepsize,
    u2_lm1
)

Integrator using Crank-Nicholson Adams-Bashforth method for (FD)

Arguments

A: A matrix

F: F matrix

tdata: temporal points

IC: initial condition

const_stepsize: whether to use a constant time step size

u2_lm1: u2 at j-2

Returns

u: state matrix

LiftAndLearn.KuramotoSivashinsky.integrate_finite_diff_model — Method

integrate_finite_diff_model(ops, tdata, IC; params...)

Integrator using Crank-Nicholson Adams-Bashforth method for (FD). This is a dispatch function for integrate_FD(A, F, tdata, IC; const_stepsize=true, u2_lm1=nothing). Using the operator struct ops instead of A and F.

Arguments

ops: Operators

tdata: temporal points

IC: initial condition

params: keyword arguments

const_stepsize: whether to use a constant time step size

Returns

u: state matrix

LiftAndLearn.KuramotoSivashinsky.integrate_pseudo_spectral_model — Method

integrate_pseudo_spectral_model(A, F, tdata, IC)

Integrator using Crank-Nicholson Adams-Bashforth method for (FFT)

Arguments

A: A matrix

F: F matrix

tdata: temporal points

IC: initial condition

Returns

u: state matrix

uhat: state matrix in the Fourier space

LiftAndLearn.KuramotoSivashinsky.integrate_spectral_galerkin_model — Method

integrate_spectral_galerkin_model(A, F, tdata, IC)

Integrator for model produced with Spectral-Galerkin method.

Arguments

A: A matrix

F: F matrix

tdata: temporal points

IC: initial condition

Returns

u: state matrix

uhat: state matrix in the Fourier space

LiftAndLearn.KuramotoSivashinsky.jacobian — Method

jacobian(ops, x)

Generate Jacobian matrix

Arguments

A: A matrix

H: H matrix

x: state

Returns

J: Jacobian matrix

LiftAndLearn.KuramotoSivashinsky.pseudo_spectral_model — Method

pseudo_spectral_model(model, μ) → A, F

Generate A, F matrices for the Kuramoto-Sivashinsky equation using the Pseudo-Spectral/Fast Fourier Transform method.

Arguments

model: Kuramoto-Sivashinsky equation model

μ: parameter value

Returns

A: A matrix

F: F matrix (take out 1.0im)

LiftAndLearn.KuramotoSivashinsky.spectral_galerkin_model — Method

spectral_galerkin_model(model, μ) → A, F

Generate A, F matrices for the Kuramoto-Sivashinsky equation using the Spectral-Galerkin method.

Arguments

model: Kuramoto-Sivashinsky equation model

μ: parameter value

Returns

A: A matrix

F: F matrix