API

API

All APIs of LiftAndLearn listed in a unstructured manner.

LiftAndLearn.EPConstraintResidual — Function

EPConstraintResidual(X, r, which_quad="H"; with_mmt=false) → ϵX, mmt

Compute the constraint residual which is the residual of the energy-preserving constraint

\[\sum \left| \hat{h}_{ijk} + \hat{h}_{jik} + \hat{h}_{kji} \right| \quad 1 \leq i,j,k \leq r\]

Arguments

X::Union{Matrix,SparseMatrixCSC}: the matrix to compute the constraint residual

r::Real: the dimension of the system

which_quad::String: the type of the quadratic operator (H or F)

with_mmt::Bool: whether to compute the moment of the constraint residual

Returns

ϵX: the constraint residual

mmt: the moment which is the sum of the constraint residual without absolute value

LiftAndLearn.EPConstraintViolation — Function

EPConstraintViolation(Data, X, which_quad="H") → viol

Compute the constraint violation which is the violation of the energy-preserving constraint

\[\sum \langle \mathbf{x}, \mathbf{H}(\mathbf{x}\otimes\mathbf{x})\rangle \quad \forall \mathbf{x} \in \mathcal{D}\]

Arguments

Data::AbstractArray: the data

X::Union{Matrix,SparseMatrixCSC}: the matrix to compute the constraint violation

which_quad::String: the type of the quadratic operator (H or F)

Returns

viol: the constraint violation

LiftAndLearn.EPHEC_Optimize — Method

EPHEC_Optimize(D::Matrix, Rt::Union{Matrix,Transpose}, dims::Dict, 
    options::Abstract_Options, IG::operators) → Ahat, Bhat, Fhat, Hhat, Nhat, Khat

Energy preserved (Hard Equality Constraint) operator inference optimization (EPHEC)

Arguments

D: data matrix

Rt: transpose of the derivative matrix

dims: important dimensions

options: options for the operator inference set by the user

IG: Initial Guesses

Returns

Inferred operators

LiftAndLearn.EPP_Optimize — Method

EPP_Optimize(D::Matrix, Rt::Union{Matrix,Transpose}, dims::Dict, 
    options::Abstract_Options, IG::operators) → Ahat, Bhat, Fhat, Hhat, Nhat, Khat

Energy preserving penalty operator inference optimization (EPP)

Arguments

D: data matrix

Rt: transpose of the derivative matrix

dims: important dimensions

options: options for the operator inference set by the user

IG: Initial Guesses

Returns

Inferred operators

LiftAndLearn.EPSIC_Optimize — Method

EPSIC_Optimize(D::Matrix, Rt::Union{Matrix,Transpose}, dims::Dict, 
    options::Abstract_Options, IG::operators) → Ahat, Bhat, Fhat, Hhat, Nhat, Khat

Energy preserved (Soft Inequality Constraint) operator inference optimization (EPSIC)

Arguments

D: data matrix

Rt: transpose of the derivative matrix

dims: important dimensions

options: options for the operator inference set by the user

IG: Initial Guesses

Returns

Inferred operators

LiftAndLearn.F2H — Method

F2H(F::Union{SparseMatrixCSC,VecOrMat}) → H

Convert the quadratic F operator into the H operator

Arguments

F: F matrix

Returns

H: H matrix

LiftAndLearn.F2Hs — Method

F2Hs(F::Union{SparseMatrixCSC,VecOrMat}) → Hs

Convert the quadratic F operator into the symmetric H operator.

This guarantees that the H operator is symmetric. The difference from F2H is that we use the elimination matrix L and the symmetric commutation matrix N to multiply the F matrix.

Arguments

F: F matrix

Returns

Hs: symmetric H matrix

LiftAndLearn.H2F — Method

H2F(H::Union{SparseMatrixCSC,VecOrMat}) → F

Convert the quadratic H operator into the F operator

Arguments

H: H matrix

Returns

F: F matrix

LiftAndLearn.H2Q — Method

H2Q(H::AbstractArray) → Q

Convert the quadratic H operator into the Q operator

Arguments

H::AbstractArray: Quadratic matrix of dimensions (n x n^2)

Returns

the Q quadratic matrix of 3-dim tensor

LiftAndLearn.LS_solve — Method

LS_solve(D::Matrix, Rt::Union{Matrix,Transpose}, Y::Matrix, Xhat_t::Union{Matrix,Transpose}, 
    dims::Dict, options::Abstract_Options) → Ahat, Bhat, Chat, Fhat, Hhat, Nhat, Khat

Solve the standard Operator Inference with/without regularization

Arguments

D::Matrix: data matrix

Rt::Union{Matrix,Transpose}: derivative data matrix (transposed)

Y::Matrix: output data matrix

Xhat_t::Union{Matrix,Transpose}: projected data matrix (transposed)

dims::Dict: dictionary including important dimensions

options::Abstract_Options: options for the operator inference set by the user

Returns

All learned operators A, B, C, F, H, N, K

LiftAndLearn.NC_Optimize — Method

NC_Optimize(D::Matrix, Rt::Union{Matrix, Transpose}, dims::Dict, 
    options::Abstract_Options, IG::operators) → Ahat, Bhat, Fhat, Hhat, Nhat, Khat

Optimization version of Standard Operator Inference (NC)

Arguments

D: data matrix

Rt: transpose of the derivative matrix

dims: important dimensions

options: options for the operator inference set by the user

IG: Initial Guesses

Returns

Inferred operators

LiftAndLearn.NC_Optimize_output — Method

NC_Optimize_output(Y::Matrix, Xt_hat::Union{Matrix, Transpose}, dims::Dict, options::Abstract_Options) → C

Output optimization for the standard operator inference (for operator C)

Arguments

Y: the output matrix

Xt_hat: the state matrix

Return

the output state matrix C

LiftAndLearn.Q2H — Method

Q2H(Q::AbstractArray) → H

Convert the quadratic Q operator into the H operator. The Q matrix is a 3-dim tensor with dimensions (n x n x n). Thus,

\[\mathbf{Q} = \begin{bmatrix} \mathbf{Q}_1 \\ \mathbf{Q}_2 \\ \vdots \\ \mathbf{Q}_n \end{bmatrix} \quad \text{where }~~ \mathbf{Q}_i \in \mathbb{R}^{n \times n}\]

Arguments

Q::AbstractArray: Quadratic matrix in the 3-dim tensor form with dimensions (n x n x n)

Returns

the H quadratic matrix

LiftAndLearn.backwardEuler — Method

backwardEuler(A, B, U, tdata, IC)

Backward Euler scheme integration.

Arguments

A: linear state operator

B: linear input operator

U: input data

tdata: time data

IC: initial condtions

Return

states: integrated states

LiftAndLearn.choose_ro — Method

choose_ro(Σ::Vector; en_low=-15) → r_all, en

Choose reduced order (ro) that preserves an acceptable energy.

Arguments

Σ::Vector: Singular value vector from the SVD of some Hankel Matrix

en_low: minimum size for energy preservation

Returns

r_all: vector of reduced orders

en: vector of energy values

LiftAndLearn.commat — Method

commat(m::Integer, n::Integer) → K

Create commutation matrix K of dimension m x n [magnus1980].

Arguments

m::Integer: row dimension of the commutation matrix

n::Integer: column dimension of the commutation matrix

Returns

K: commutation matrix

LiftAndLearn.commat — Method

commat(m::Integer) → K

Dispatch for the commutation matrix of dimensions (m, m)

Arguments

m::Integer: row and column dimension of the commutation matrix

Returns

K: commutation matrix

LiftAndLearn.compError — Method

compError(Xf, Yf, Xint, Yint, Xinf, Yinf, Vr) → PE, ISE, IOE, OSE, OOE

Compute all projection, state, and output errors

Arguments

Xf: reference state data

Yf: reference output data

Xint: intrusive model state data

Yint: intrusive model output data

Xinf: inferred model state data

Xint: inferrred model output data

Vr: POD basis

Return

PE: projection error

ISE: intrusive state error

IOE: intrusive output error

OSE: operator inference state error

OOE: operator inference output error

LiftAndLearn.compOutputError — Method

compOutputError(Yf, Y) → OE

Compute output error

Arguments

Yf: reference output data

Y: testing output data

Return

OE: output error

LiftAndLearn.compProjError — Method

compProjError(Xf, Vr) → PE

Compute the projection error

Arguments

Xf: reference state data

Vr: POD basis

Return

PE: projection error

LiftAndLearn.compStateError — Method

compStateError(Xf, X, Vr) → SE

Compute the state error

Arguments

Xf: reference state data

X: testing state data

Vr: POD basis

Return

SE: state error

LiftAndLearn.crankNicolson — Method

crankNicolson(A, B, U, tdata, IC)

Crank-Nicolson scheme

Arguments

A: linear state operator

B: linear input operator

U: input data

tdata: time data

IC: initial condtions

Return

states: integrated states

LiftAndLearn.delta — Method

delta(v::Int, w::Int) → Float64

Another auxiliary function for the F matrix

Arguments

v: first index

w: second index

Returns

coefficient of 1.0 or 0.5

LiftAndLearn.dtApprox — Method

dtApprox(X::VecOrMat, options::Abstract_Options) → dXdt, idx

Approximating the derivative values of the data with different integration schemes

Arguments

X::VecOrMat: data matrix

options::Abstract_Options: operator inference options

Returns

dXdt: derivative data

idx: index for the specific integration scheme (important for later use)

LiftAndLearn.dupmat — Method

dupmat(n::Integer) → D

Create duplication matrix D of dimension n [magnus1980].

Arguments

n: dimension of the duplication matrix

Returns

D: duplication matrix

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  3
 3  4

julia> D = LnL.dupmat(2)
4×3 SparseArrays.SparseMatrixCSC{Float64, Int64} with 4 stored entries:
 1.0   ⋅    ⋅ 
  ⋅   1.0   ⋅
  ⋅   1.0   ⋅
  ⋅    ⋅   1.0

julia> D * LnL.vech(A)
4-element Vector{Float64}:
 1.0
 3.0
 3.0
 4.0

julia> a = vec(A)
4-element Vector{Int64}:
 1
 3
 3
 4

References

[magnus1980] J. R. Magnus and H. Neudecker, “The Elimination Matrix: Some Lemmas and Applications,” SIAM. J. on Algebraic and Discrete Methods, vol. 1, no. 4, pp. 422–449, Dec. 1980, doi: 10.1137/0601049.

LiftAndLearn.elimat — Method

elimat(m::Integer) → L

Create elimination matrix L of dimension m [magnus1980].

Arguments

m: dimension of the target matrix

Return

L: elimination matrix

LiftAndLearn.extractF — Method

extractF(F::Union{SparseMatrixCSC,VecOrMat}, r::Int) → F

Extracting the F matrix for POD basis of dimensions (N, r)

Arguments

F: F matrix

r: reduced order

Returns

extracted F matrix

LiftAndLearn.extractH — Method

extractH(H::Union{SparseMatrixCSC,VecOrMat}, r::Int) → H

Extracting the H matrix for POD basis of dimensions (N, r)

Arguments

H: H matrix

r: reduced order

Returns

extracted H matrix

LiftAndLearn.fidx — Method

fidx(n::Int, j::Int, k::Int) → Int

Auxiliary function for the F matrix indexing.

Arguments

n: row dimension of the F matrix

j: row index

k: col index

Returns

index corresponding to the F matrix

LiftAndLearn.forwardEuler — Method

forwardEuler(f, g, tdata, IC)

Forward euler scheme [dispatch for f(x,u) and u = g(u,t)]

Arguments

f: Xdot = f(x,g(u,t)) right-hand-side of the dynamics

g: Xdot = f(x,g(u,t)) input function g(u,t)

tdata: time data

IC: initial conditions

Return

states: integrated states

LiftAndLearn.forwardEuler — Method

forwardEuler(f, U, tdata, IC)

Forward euler scheme [dispatch for f(x,u) and u-input as U-matrix]

Arguments

f: Xdot = f(x,U) right-hand-side of the dynamics

U: Xdot = f(x,U) input data U

tdata: time data

IC: initial conditions

Return

states: integrated states

LiftAndLearn.forwardEuler — Method

forwardEuler(A, B, U, tdata, IC)

Forward euler scheme integration.

Arguments

A: linear state operator

B: linear input operator

U: input vector/matrix

tdata: time data

IC: initial conditions

Return

states: integrated states

LiftAndLearn.getDataMat — Method

getDataMat(Xhat::Matrix, Xhat_t::Union{Matrix,Transpose}, U::Matrix,
    dims::Dict, options::Abstract_Options) → D

Get the data matrix for the regression problem

Arguments

Xhat::Matrix: projected data matrix

Xhat_t::Union{Matrix,Transpose}: projected data matrix (transposed)

U::Matrix: input data matrix

dims::Dict: dictionary including important dimensions

options::Abstract_Options: options for the operator inference set by the user

Returns

D: data matrix for the regression problem

LiftAndLearn.inferOp — Function

inferOp(X::Matrix, U::Matrix, Y::VecOrMat, Vn::Matrix,
    options::Abstract_Options, IG::operators=operators()) → op::operators

Infer the operators without derivative data (dispatch)

Arguments

X::Matrix: state data matrix

U::Matrix: input data matrix

Y::VecOrMat: output data matix

Vn::Matrix: POD basis

options::Abstract_Options: options for the operator inference defined by the user

IG::operators: initial guesses for optimization

Returns

op::operators: inferred operators

LiftAndLearn.inferOp — Function

inferOp(X::Matrix, U::Matrix, Y::VecOrMat, Vn::Matrix, R::Matrix,
    options::Abstract_Options, IG::operators=operators()) → op::operators

Infer the operators with derivative data given

Arguments

X::Matrix: state data matrix

U::Matrix: input data matrix

Y::VecOrMat: output data matix

Vn::Matrix: POD basis

R::Matrix: derivative data matrix

options::Abstract_Options: options for the operator inference defined by the user

IG::operators: initial guesses for optimization

Returns

op::operators: inferred operators

LiftAndLearn.inferOp — Function

inferOp(X::Matrix, U::Matrix, Y::VecOrMat, Vn::Matrix,
    full_op::operators, options::Abstract_Options, IG::operators=operators()) → op::operators

Infer the operators with reprojection method (dispatch)

Arguments

X::Matrix: state data matrix

U::Matrix: input data matrix

Y::VecOrMat: output data matix

Vn::Matrix: POD basis

full_op::operators: full order model operators

options::Abstract_Options: options for the operator inference defined by the user

IG::operators: initial guesses for optimization

Returns

op::operators: inferred operators

LiftAndLearn.inferOp — Function

inferOp(W::Matrix, U::Matrix, Y::VecOrMat, Vn::Union{Matrix,BlockDiagonal},
    lm::lifting, full_op::operators, options::Abstract_Options, 
    IG::operators=operators()) → op::operators

Infer the operators for Lift And Learn for reprojected data (dispatch)

Arguments

W::Matrix: state data matrix

U::Matrix: input data matrix

Y::VecOrMat: output data matix

Vn::Union{Matrix,BlockDiagonal}: POD basis

lm::lifting: struct of the lift map

full_op::operators: full order model operators

options::Abstract_Options: options for the operator inference defined by the user

IG::operators: initial guesses for optimization

Returns

op::operators: inferred operators

LiftAndLearn.insert2F — Method

insertF(Fi::Union{SparseMatrixCSC,VecOrMat}, N::Int) → F

Inserting values into the F matrix for higher dimensions

Arguments

Fi: F matrix to insert

N: the larger order

Returns

inserted F matrix

LiftAndLearn.insert2H — Method

insertH(Hi::Union{SparseMatrixCSC,VecOrMat}, N::Int) → H

Inserting values into the H matrix for higher dimensions

Arguments

Hi: H matrix to insert

N: the larger order

Returns

inserted H matrix

LiftAndLearn.insert2bilin — Method

insert2bilin(X::Union{SparseMatrixCSC,VecOrMat}, N::Int, p::Int) → BL

Inserting the values into the bilinear matrix (N) for higher dimensions

Arguments

X: bilinear matrix to insert

N: the larger order

Returns

Inserted bilinear matrix

LiftAndLearn.insert2randF — Method

insert2randF(Fi::Union{SparseMatrixCSC,VecOrMat}, N::Int) → F

Inserting values into the F matrix for higher dimensions

Arguments

Fi: F matrix to insert

N: the larger order

Returns

inserted F matrix

LiftAndLearn.intrusiveMR — Method

intrusiveMR(op, Vr, options)

Perform intrusive model reduction using Proper Orthogonal Decomposition (POD)

Arguments

op: operators of the target system (A, B, C, F/H, N, K)

Vr: POD basis

options: options for the operator inference

Return

op_new: new operator projected onto the basis

LiftAndLearn.invec — Method

invec(r::AbstractArray, m::Int, n::Int) → r

Inverse vectorization.

Arguments

r::AbstractArray: the input vector

m::Int: the row dimension

n::Int: the column dimension

Returns

the inverse vectorized matrix

LiftAndLearn.kronMatStates — Method

kronMatStates(Xmat::Union{SparseMatrixCSC,VecOrMat}) → Xkron

Generate the kronecker product state values (corresponding to the H matrix) for a matrix form state data

Arguments

Xmat: state snapshot matrix

Returns

kronecker product state snapshot matrix

LiftAndLearn.liftedBasis — Method

liftedBasis(W, Nl, gp, ro) → Vr

Create the block-diagonal POD basis for the new lifted system data

Arguments

w: lifted data matrix

Nl: number of variables of the lifted state dynamics

gp: number of grid points for each variable

ro: vector of the reduced orders for each basis

Return

Vr: block diagonal POD basis

LiftAndLearn.nommat — Method

nommat(m::Integer, n::Integer) → N

Create symmetric commutation matrix N of dimension m x n [magnus1980].

Arguments

m::Integer: row dimension of the commutation matrix

n::Integer: column dimension of the commutation matrix

Returns

N: symmetric commutation matrix

LiftAndLearn.nommat — Method

nommat(m::Integer) → N

Dispatch for the symmetric commutation matrix of dimensions (m, m)

Arguments

m::Integer: row and column dimension of the commutation matrix

Returns

N: symmetric commutation matrix

LiftAndLearn.reproject — Method

inferOp(W::Matrix, U::Matrix, Y::VecOrMat, Vn::Union{Matrix,BlockDiagonal},
    lm::lifting, options::Abstract_Options, IG::operators=operators()) → op::operators

Reprojecting the lifted data

Arguments

Xhat::Matrix: state data matrix projected onto the basis

V::Union{VecOrMat,BlockDiagonal}: POD basis

U::VecOrMat: input data matrix

dims::Dict: dictionary including important dimensions

lm::lifting: struct of the lift map

op::operators: full order model operators

options::Abstract_Options: options for the operator inference defined by the user

Returns

Rhat::Matrix: R matrix (transposed) for the regression problem

LiftAndLearn.reproject — Method

inferOp(W::Matrix, U::Matrix, Y::VecOrMat, Vn::Union{Matrix,BlockDiagonal},
    lm::lifting, options::Abstract_Options, IG::operators=operators()) → op::operators

Reprojecting the data to minimize the error affected by the missing orders of the POD basis

Arguments

Xhat::Matrix: state data matrix projected onto the basis

V::Union{VecOrMat,BlockDiagonal}: POD basis

U::VecOrMat: input data matrix

dims::Dict: dictionary including important dimensions

op::operators: full order model operators

options::Abstract_Options: options for the operator inference defined by the user

Return

Rhat::Matrix: R matrix (transposed) for the regression problem

LiftAndLearn.run_optimizer — Function

run_optimizer(D::AbstractArray, Rt::AbstractArray, Y::AbstractArray,
    Xhat_t::AbstractArray, dims::Dict, options::Abstract_Options,
    IG::operators=operators()) → op::operators

Run the optimizer of choice.

Arguments

D::AbstractArray: data matrix

Rt::AbstractArray: derivative data matrix (transposed)

Y::AbstractArray: output data matrix

Xhat_t::AbstractArray: projected data matrix (transposed)

dims::Dict: dictionary including important dimensions

options::Abstract_Options: options for the operator inference set by the user

IG::operators: initial guesses for optimization

Returns

op::operators: All learned operators

LiftAndLearn.semiImplicitEuler — Method

semiImplicitEuler(A, B, F, U, tdata, IC)

Semi-Implicit Euler scheme

Arguments

A: linear state operator

B: linear input operator

F: quadratic state operator

U: input data

tdata: time data

IC: initial condtions

Return

states: integrated states

LiftAndLearn.semiImplicitEuler — Method

semiImplicitEuler(A, B, F_or_H, U, tdata, IC, options)

Semi-Implicit Euler scheme (dispatch)

Arguments

A: linear state operator

B: linear input operator

F_or_H: quadratic state operator (F or H)

U: input data

tdata: time data

IC: initial condtions

Return

states: integrated states

LiftAndLearn.squareMatStates — Method

squareMatStates(Xmat::Union{SparseMatrixCSC,VecOrMat}) → Xsq

Generate the x^[2] squared state values (corresponding to the F matrix) for a snapshot data matrix

Arguments

Xmat: state snapshot matrix

Returns

squared state snapshot matrix

LiftAndLearn.tikhonov — Method

tikhonov(b::AbstractArray, A::AbstractArray, Γ::AbstractMatrix, tol::Real;
    flag::Bool=false) → x

Tikhonov regression

Arguments

b::AbstractArray: right hand side of the regression problem

A::AbstractArray: left hand side of the regression problem

Γ::AbstractMatrix: Tikhonov matrix

tol::Real: tolerance for the singular values

flag::Bool: flag for the tolerance

Returns

regression solution

LiftAndLearn.tikhonovMatrix! — Method

tikhonovMatrix!(Γ::AbstractArray, dims::Dict, options::Abstract_Options)

Construct the Tikhonov matrix

Arguments

Γ::AbstractArray: Tikhonov matrix (pass by reference)

dims::Dict: dictionary including important dimensions

options::Abstract_Options: options for the operator inference set by the user

Returns

Γ: Tikhonov matrix (pass by reference)

LiftAndLearn.vech — Method

vech(A::AbstractMatrix{T}) → v

Half-vectorization operation. For example half-vectorzation of

\[A = \begin{bmatrix} a_{11} & a_{12} \\ a_{21} & a_{22} \end{bmatrix}\]

becomes

\[v = \begin{bmatrix} a_{11} \\ a_{21} \\ a_{22} \end{bmatrix}\]

Arguments

A: matrix to half-vectorize

Returns

v: half-vectorized form