Lifting

This page gives an explanation for the mathematical concept of lifting and provides example code for its implementation.

Lift Map Structure

LiftAndLearn.lifting — Type

struct lifting

Lifting map structure.

Fields

N: number of variables of the original nonlinear dynamics

Nl: number of variables of the lifted system

lift_funcs: array of lifting transformation functions

map: function to map the data to the new mapped states including original states

mapNL: function to map the data to only the additional lifted states

Tip

For a simple pendulum we have

\[\begin{bmatrix} \dot{x}_1 \\ \dot{x}_2 \end{bmatrix} = \begin{bmatrix} x_2 \\ -\frac{g}{l} \sin(x_1) \end{bmatrix}\]

The lifted system becomes

\[\begin{bmatrix} \dot{x}_1 \\ \dot{x}_2 \\ \dot{x}_3 \\ \dot{x}_4 \end{bmatrix} = \begin{bmatrix} x_2 \\ -\frac{g}{l} x_3 \\ x_2 x_4 \\ -x_2 x_3 \end{bmatrix}\]

when $x_3 = \sin(x_1)$ and $x_4 = \cos(x_1)$. Which if coded, would look like this:

lifter = LnL.lifting(2, 4, [x -> sin.(x[1]), x -> cos.(x[1])])

Construct Lifted POD Basis from Data

LiftAndLearn.liftedBasis — Function

liftedBasis(W, Nl, gp, ro) → Vr

Create the block-diagonal POD basis for the new lifted system data

Arguments

w: lifted data matrix

Nl: number of variables of the lifted state dynamics

gp: number of grid points for each variable

ro: vector of the reduced orders for each basis

Return

Vr: block diagonal POD basis

An example implementation would be:

using LiftAndLearn
using Random
LnL = LiftAndLearn
W = round.(rand(30,100), digits=4)
LnL.liftedBasis(W, 3, 10, [2,3,4])

30×9 BlockDiagonals.BlockDiagonal{Float64, Matrix{Float64}}:
 -0.324598   0.13795    0.0  0.0  0.0  …   0.0         0.0         0.0
 -0.284986  -0.233793   0.0  0.0  0.0      0.0         0.0         0.0
 -0.326167  -0.321824   0.0  0.0  0.0      0.0         0.0         0.0
 -0.333515   0.19595    0.0  0.0  0.0      0.0         0.0         0.0
 -0.307225   0.318364   0.0  0.0  0.0      0.0         0.0         0.0
 -0.313071  -0.347533   0.0  0.0  0.0  …   0.0         0.0         0.0
 -0.323371  -0.439115   0.0  0.0  0.0      0.0         0.0         0.0
 -0.289107   0.600099   0.0  0.0  0.0      0.0         0.0         0.0
 -0.310506   0.0875963  0.0  0.0  0.0      0.0         0.0         0.0
 -0.344725   0.0399267  0.0  0.0  0.0      0.0         0.0         0.0
  ⋮                                    ⋱                          
  0.0        0.0        0.0  0.0  0.0      0.128036   -0.372317   -0.122313
  0.0        0.0        0.0  0.0  0.0      0.316658    0.459889    0.284859
  0.0        0.0        0.0  0.0  0.0     -0.375552    0.116996    0.337322
  0.0        0.0        0.0  0.0  0.0     -0.13263     0.0847683   0.527182
  0.0        0.0        0.0  0.0  0.0  …   0.48062     0.0673413  -0.215562
  0.0        0.0        0.0  0.0  0.0     -0.293347    0.435585   -0.44852
  0.0        0.0        0.0  0.0  0.0     -0.412779   -0.48576     0.123441
  0.0        0.0        0.0  0.0  0.0     -0.234612    0.280192   -0.219806
  0.0        0.0        0.0  0.0  0.0     -0.0787597  -0.248066   -0.424899