Utilities

System operators

The struct operators provides a convenient way to organize and store the full and reduced operators.

LiftAndLearn.operators — Type

mutable struct operators

Organize the operators of the system in a structure. The operators currently supported are up to second order.

Fields

A: linear state operator

B: linear input operator

C: linear output operator

F: quadratic state operator with no redundancy

H: quadratic state operator with redundancy

Q: quadratic state operator with redundancy in 3-dim tensor form

G: cubic state operator with redundancy

E: cubic state operator with no redundancy

K: constant operator

N: bilinear (state-input) operator

f: nonlinear function operator f(x,u)

Linear Algebra

Many of the utility functions revolve around the properties of the following concepts

Kronecker Products $\otimes$

Vectorization $\text{vec}(\cdot)$

Half-vectorization $\text{vech}(\cdot)$

The utility functions are based on the following key references

(1) Brewer, J. W. Kronecker Products and Matrix Calculus in System Theory IEEE Transactions on Circuits and Systems, 25(9) 772-781, 1978.

@article{Brewer1978,
  title = {Kronecker products and matrix calculus in system theory},
  volume = {25},
  ISSN = {0098-4094},
  url = {http://dx.doi.org/10.1109/TCS.1978.1084534},
  DOI = {10.1109/tcs.1978.1084534},
  number = {9},
  journal = {IEEE Transactions on Circuits and Systems},
  publisher = {Institute of Electrical and Electronics Engineers (IEEE)},
  author = {Brewer,  J.},
  year = {1978},
  month = sep,
  pages = {772–781}
}

(2) Magnus, J. R. and Neudecker, H. The Elimination Matrix: Some Lemmas and Applications. SIAM Journal on Algebraic Discrete Methods, 1(4) 422-449, 1980-12.

@article{Magnus1980,
  title = {The Elimination Matrix: Some Lemmas and Applications},
  volume = {1},
  ISSN = {2168-345X},
  url = {http://dx.doi.org/10.1137/0601049},
  DOI = {10.1137/0601049},
  number = {4},
  journal = {SIAM Journal on Algebraic Discrete Methods},
  publisher = {Society for Industrial & Applied Mathematics (SIAM)},
  author = {Magnus,  Jan R. and Neudecker,  H.},
  year = {1980},
  month = dec,
  pages = {422–449}
}

Utility APIs

LiftAndLearn.vech — Function

vech(A::AbstractMatrix{T}) → v

Half-vectorization operation. For example half-vectorzation of

\[A = \begin{bmatrix} a_{11} & a_{12} \\ a_{21} & a_{22} \end{bmatrix}\]

becomes

\[v = \begin{bmatrix} a_{11} \\ a_{21} \\ a_{22} \end{bmatrix}\]

Arguments

A: matrix to half-vectorize

Returns

v: half-vectorized form

LiftAndLearn.:⊘ — Function

⊘(x::AbstractVector{T}, y::AbstractVector{T}) where T

Unique Kronecker product operation

Arguments

x::AbstractVector{T}: vector to perform the unique Kronecker product

y::AbstractVector{T}: vector to perform the unique Kronecker product

Returns

unique Kronecker product

LiftAndLearn.invec — Function

invec(r::AbstractArray, m::Int, n::Int) → r

Inverse vectorization.

Arguments

r::AbstractArray: the input vector

m::Int: the row dimension

n::Int: the column dimension

Returns

the inverse vectorized matrix

LiftAndLearn.dupmat — Function

dupmat(n::Integer) → D

Create duplication matrix D of dimension n [magnus1980].

Arguments

n: dimension of the duplication matrix

Returns

D: duplication matrix

Examples

julia> A = [1 2; 3 4]
2×2 Matrix{Int64}:
 1  3
 3  4

julia> D = LnL.dupmat(2)
4×3 SparseArrays.SparseMatrixCSC{Float64, Int64} with 4 stored entries:
 1.0   ⋅    ⋅ 
  ⋅   1.0   ⋅
  ⋅   1.0   ⋅
  ⋅    ⋅   1.0

julia> D * LnL.vech(A)
4-element Vector{Float64}:
 1.0
 3.0
 3.0
 4.0

julia> a = vec(A)
4-element Vector{Int64}:
 1
 3
 3
 4

References

[magnus1980] J. R. Magnus and H. Neudecker, “The Elimination Matrix: Some Lemmas and Applications,” SIAM. J. on Algebraic and Discrete Methods, vol. 1, no. 4, pp. 422–449, Dec. 1980, doi: 10.1137/0601049.

LiftAndLearn.elimat — Function

elimat(m::Integer) → L

Create elimination matrix L of dimension m [magnus1980].

Arguments

m: dimension of the target matrix

Return

L: elimination matrix

LiftAndLearn.commat — Function

commat(m::Integer, n::Integer) → K

Create commutation matrix K of dimension m x n [magnus1980].

Arguments

m::Integer: row dimension of the commutation matrix

n::Integer: column dimension of the commutation matrix

Returns

K: commutation matrix

commat(m::Integer) → K

Dispatch for the commutation matrix of dimensions (m, m)

Arguments

m::Integer: row and column dimension of the commutation matrix

Returns

K: commutation matrix

LiftAndLearn.symmtzrmat — Function

symmtzrmat(m::Integer, n::Integer) → N

Create symmetrizer (or symmetric commutation) matrix N of dimension m x n [magnus1980].

Arguments

m::Integer: row dimension of the commutation matrix

n::Integer: column dimension of the commutation matrix

Returns

N: symmetrizer (symmetric commutation) matrix

symmtzrmat(m::Integer) → N

Dispatch for the symmetric commutation matrix of dimensions (m, m)

Arguments

m::Integer: row and column dimension of the commutation matrix

Returns

N: symmetric commutation matrix

LiftAndLearn.F2H — Function

F2H(F::Union{SparseMatrixCSC,VecOrMat}) → H

Convert the quadratic F operator into the H operator

Arguments

F: F matrix

Returns

H: H matrix

LiftAndLearn.H2F — Function

H2F(H::Union{SparseMatrixCSC,VecOrMat}) → F

Convert the quadratic H operator into the F operator

Arguments

H: H matrix

Returns

F: F matrix

LiftAndLearn.F2Hs — Function

F2Hs(F::Union{SparseMatrixCSC,VecOrMat}) → Hs

Convert the quadratic F operator into the symmetric H operator.

This guarantees that the H operator is symmetric. The difference from F2H is that we use the elimination matrix L and the symmetric commutation matrix N to multiply the F matrix.

Arguments

F: F matrix

Returns

Hs: symmetric H matrix

LiftAndLearn.Q2H — Function

Q2H(Q::AbstractArray) → H

Convert the quadratic Q operator into the H operator. The Q matrix is a 3-dim tensor with dimensions (n x n x n). Thus,

\[\mathbf{Q} = \begin{bmatrix} \mathbf{Q}_1 \\ \mathbf{Q}_2 \\ \vdots \\ \mathbf{Q}_n \end{bmatrix} \quad \text{where }~~ \mathbf{Q}_i \in \mathbb{R}^{n \times n}\]

Arguments

Q::AbstractArray: Quadratic matrix in the 3-dim tensor form with dimensions (n x n x n)

Returns

the H quadratic matrix

LiftAndLearn.H2Q — Function

H2Q(H::AbstractArray) → Q

Convert the quadratic H operator into the Q operator

Arguments

H::AbstractArray: Quadratic matrix of dimensions (n x n^2)

Returns

the Q quadratic matrix of 3-dim tensor

LiftAndLearn.squareMatStates — Function

squareMatStates(Xmat::Union{SparseMatrixCSC,VecOrMat}) → Xsq

Generate the x^[2] squared state values (corresponding to the F matrix) for a snapshot data matrix

Arguments

Xmat: state snapshot matrix

Returns

squared state snapshot matrix

LiftAndLearn.kronMatStates — Function

kronMatStates(Xmat::Union{SparseMatrixCSC,VecOrMat}) → Xkron

Generate the kronecker product state values (corresponding to the H matrix) for a matrix form state data

Arguments

Xmat: state snapshot matrix

Returns

kronecker product state snapshot matrix

LiftAndLearn.extractF — Function

extractF(F::Union{SparseMatrixCSC,VecOrMat}, r::Int) → F

Extracting the F matrix for POD basis of dimensions (N, r)

Arguments

F: F matrix

r: reduced order

Returns

extracted F matrix

LiftAndLearn.insert2F — Function

insertF(Fi::Union{SparseMatrixCSC,VecOrMat}, N::Int) → F

Inserting values into the F matrix for higher dimensions

Arguments

Fi: F matrix to insert

N: the larger order

Returns

inserted F matrix

LiftAndLearn.insert2randF — Function

insert2randF(Fi::Union{SparseMatrixCSC,VecOrMat}, N::Int) → F

Inserting values into the F matrix for higher dimensions

Arguments

Fi: F matrix to insert

N: the larger order

Returns

inserted F matrix

LiftAndLearn.extractH — Function

extractH(H::Union{SparseMatrixCSC,VecOrMat}, r::Int) → H

Extracting the H matrix for POD basis of dimensions (N, r)

Arguments

H: H matrix

r: reduced order

Returns

extracted H matrix

LiftAndLearn.insert2H — Function

insertH(Hi::Union{SparseMatrixCSC,VecOrMat}, N::Int) → H

Inserting values into the H matrix for higher dimensions

Arguments

Hi: H matrix to insert

N: the larger order

Returns

inserted H matrix

LiftAndLearn.insert2bilin — Function

insert2bilin(X::Union{SparseMatrixCSC,VecOrMat}, N::Int, p::Int) → BL

Inserting the values into the bilinear matrix (N) for higher dimensions

Arguments

X: bilinear matrix to insert

N: the larger order

Returns

Inserted bilinear matrix