Fitzhugh-Nagumo Equation

LiftAndLearn.FHN — Module

Fitzhugh-Nagumo PDE model

LiftAndLearn.FHN.Abstract_Models — Type

Abstract_Models

Abstract type for the models.

LiftAndLearn.FHN.fhn — Type

mutable struct fhn <: LiftAndLearn.FHN.Abstract_Models

Fitzhugh-Nagumo PDE model

\[\begin{aligned} \frac{\partial u}{\partial t} &= \epsilon^2\frac{\partial^2 u}{\partial x^2} + u(u-0.1)(1-u) - v + g \\ \frac{\partial v}{\partial t} &= hu + \gamma v + g \end{aligned}\]

where $u$ and $v$ are the state variables, $g$ is the control input, and $h$, $\gamma$, and $\epsilon$ are the parameters. Specifically, for this problem we assume the control input to begin

\[g(t) = \alpha t^3 \exp(-\beta t)\]

where $\alpha$ and $\beta$ are the parameters that are going to be varied for training.

Fields

Ω::Vector{Float64}: spatial domain

T::Vector{Float64}: temporal domain

αD::Vector{Float64}: alpha parameter domain

βD::Vector{Float64}: beta parameter domain

Δx::Float64: spatial grid size

Δt::Float64: temporal step size

Ubc::Matrix{Float64}: boundary condition (input)

ICx::Matrix{Float64}: initial condition for the original states

ICw::Matrix{Float64}: initial condition for the lifted states

x::Vector{Float64}: spatial grid points

t::Vector{Float64}: temporal points

Xdim::Int64: spatial dimension

Tdim::Int64: temporal dimension

FOM::Function: function to generate the full order operators

generateFHNmatrices::Function: function to generate the full order operators for the intrusive model operators

LiftAndLearn.FHN.fhn — Method

fhn(Ω, T, αD, βD, Δx, Δt) → fhn

Fitzhugh-Nagumo PDE model

Arguments

Ω::Vector{Float64}: spatial domain

T::Vector{Float64}: temporal domain

αD::Vector{Float64}: alpha parameter domain

βD::Vector{Float64}: beta parameter domain

Δx::Float64: spatial grid size

Δt::Float64: temporal step size

Returns

fhn: Fitzhugh-Nagumo PDE model

LiftAndLearn.FHN.FOM — Method

FOM(k, l) → A, B, C, K, f

Create the full order operators with the nonlinear operator expressed as f(x).

Arguments

k::Int64: number of spatial grid points

l::Float64: spatial domain length

Returns

A::SparseMatrixCSC{Float64,Int64}: A matrix

B::SparseMatrixCSC{Float64,Int64}: B matrix

C::SparseMatrixCSC{Float64,Int64}: C matrix

K::SparseMatrixCSC{Float64,Int64}: K matrix

f::Function: nonlinear operator

LiftAndLearn.FHN.generateFHNmatrices — Method

generateFHNmatrices(k, l) → A, B, C, H, N, K

Generate the full order operators used for the intrusive model operators

Arguments

k::Int64: number of spatial grid points

l::Float64: spatial domain length

Returns

A::SparseMatrixCSC{Float64,Int64}: A matrix

B::SparseMatrixCSC{Float64,Int64}: B matrix

C::SparseMatrixCSC{Float64,Int64}: C matrix

H::SparseMatrixCSC{Float64,Int64}: H matrix

N::SparseMatrixCSC{Float64,Int64}: N matrix

K::SparseMatrixCSC{Float64,Int64}: K matrix