Kuramoto-Sivashinsky Equation

LiftAndLearn.KS — Module

Kuramoto-Sivashinsky equation PDE model

LiftAndLearn.KS.ks — Type

mutable struct ks <: LiftAndLearn.Abstract_Model

Kuramoto-Sivashinsky equation PDE model

\[\frac{\partial u}{\partial t} = -\mu\frac{\partial^4 u}{\partial x^4} - \frac{\partial^2 u}{\partial x^2} - u\frac{\partial u}{\partial x}\]

where $u$ is the state variable and $\mu$ is the viscosity coefficient.

Fields

Omega::Vector{Float64}: spatial domain

T::Vector{Float64}: temporal domain

D::Vector{Float64}: parameter domain

nx::Float64: number of spatial grid points

Δx::Float64: spatial grid size

Δt::Float64: temporal step size

IC::VecOrMat{Float64}: initial condition

x::Vector{Float64}: spatial grid points

t::Vector{Float64}: temporal points

k::Vector{Float64}: Fourier modes

μs::Union{Vector{Float64},Float64}: parameter vector

Xdim::Int64: spatial dimension

Tdim::Int64: temporal dimension

Pdim::Int64: parameter dimension

type::String: model type

model_PS::Function: model using Pseudo-Spectral Method/Fast Fourier Transform

model_PS_ew::Function: model using Pseudo-Spectral Method/Fast Fourier Transform (element-wise)

model_SG::Function: model using Spectral-Galerkin Method

model_FD::Function: model using Finite Difference

integrate_FD::Function: integrator using Crank-Nicholson Adams-Bashforth method

integrate_PS::Function: integrator using Crank-Nicholson Adams-Bashforth method in the Fourier space

integrate_PS_ew::Function: integrator using Crank-Nicholson Adams-Bashforth method in the Fourier space (element-wise)

integrate_SG::Function: integrator for second method of Fourier Transform without FFT

LiftAndLearn.KS.ks — Method

ks(Omega, T, D, nx, Δt, Pdim, type) → ks

Kuramoto-Sivashinsky equation PDE model constructor

Arguments

Omega::Vector{Float64}: spatial domain

T::Vector{Float64}: temporal domain

D::Vector{Float64}: parameter domain

nx::Float64: number of spatial grid points

Δt::Float64: temporal step size

Pdim::Int64: parameter dimension

Returns

ks: Kuramoto-Sivashinsky equation PDE model

LiftAndLearn.KS.integrate_FD — Method

integrate_FD(A, F, tdata, IC; const_stepsize=true, u2_lm1=nothing) → u

Integrator using Crank-Nicholson Adams-Bashforth method for (FD)

Arguments

A: A matrix

F: F matrix

tdata: temporal points

IC: initial condition

const_stepsize: whether to use a constant time step size

u2_lm1: u2 at j-2

Returns

u: state matrix

LiftAndLearn.KS.integrate_FD — Method

integrate_FD(ops, tdata, IC; const_stepsize=true, u2_lm1=nothing) → u

Integrator using Crank-Nicholson Adams-Bashforth method for (FD). This is a dispatch function for integrate_FD(A, F, tdata, IC; const_stepsize=true, u2_lm1=nothing). Using the operator struct ops instead of A and F.

Arguments

ops: operators

tdata: temporal points

IC: initial condition

params: keyword arguments

const_stepsize: whether to use a constant time step size

Returns

u: state matrix

LiftAndLearn.KS.integrate_PS — Method

integrate_PS(A, F, tdata, IC) → u, uhat

Integrator using Crank-Nicholson Adams-Bashforth method for (FFT)

Arguments

A: A matrix

F: F matrix

tdata: temporal points

IC: initial condition

Returns

u: state matrix

uhat: state matrix in the Fourier space

LiftAndLearn.KS.integrate_PS_ew — Method

integrate_PS_ew(A, F, tdata, IC) → u, uhat

Integrator using Crank-Nicholson Adams-Bashforth method for (FFT) (element-wise)

Arguments

A: A matrix

F: F matrix

tdata: temporal points

IC: initial condition

Returns

u: state matrix

uhat: state matrix in the Fourier space

LiftAndLearn.KS.integrate_SG — Method

integrate_SG(A, F, tdata, IC) → u, uhat

Integrator for model produced with Spectral-Galerkin method.

Arguments

A: A matrix

F: F matrix

tdata: temporal points

IC: initial condition

Returns

u: state matrix

uhat: state matrix in the Fourier space

LiftAndLearn.KS.jacob — Method

jacob(A, H, n, x) → Jacobian matrix

Generate Jacobian matrix

Arguments

A: A matrix

H: H matrix

x: state

Returns

J: Jacobian matrix

LiftAndLearn.KS.model_FD — Method

model_FD(model, μ) → A, F

Generate A, F matrices for the Kuramoto-Sivashinsky equation using the Finite Difference method.

Arguments

model: Kuramoto-Sivashinsky equation model

μ: parameter value

Returns

A: A matrix

F: F matrix

LiftAndLearn.KS.model_PS — Method

model_PS(model, μ) → A, F

Generate A, F matrices for the Kuramoto-Sivashinsky equation using the Pseudo-Spectral/Fast Fourier Transform method.

Arguments

model: Kuramoto-Sivashinsky equation model

μ: parameter value

Returns

A: A matrix

F: F matrix (take out 1.0im)

LiftAndLearn.KS.model_PS_ew — Method

model_PS_ew(model, μ) → A, F

Generate A, F matrices for the Kuramoto-Sivashinsky equation using the Fast Fourier Transform method (element-wise).

Arguments

model: Kuramoto-Sivashinsky equation model

μ: parameter value

Returns

A: A matrix

F: F matrix

LiftAndLearn.KS.model_SG — Method

model_SG(model, μ) → A, F

Generate A, F matrices for the Kuramoto-Sivashinsky equation using the Spectral-Galerkin method.

Arguments

model: Kuramoto-Sivashinsky equation model

μ: parameter value

Returns

A: A matrix

F: F matrix